remove transaction api
